% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{doping}
\alias{doping}
\title{ECWM questions on substance use}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1211 rows and 12 columns.
}
\usage{
doping
}
\description{
A data set containing the answers and randomization probabilities of four ECWM questions,
and four covariates:
\describe{
\item{Q1}{Have you ever been drug tested, for any reason?}
\item{Q2}{Have you ever used nutritional or herbal supplements for reasons other than health?}
\item{Q3}{Have you ever used controlled psychoactive drugs?}
\item{Q4}{Have you ever used controlled sport drugs without medical need, to enhance your athletic
performance or appearance?}
\item{P1}{probability of YES answer to unrelated question on Q1}
\item{P2}{probability of YES answer to unrelated question on Q2}
\item{P3}{probability of YES answer to unrelated question on Q3}
\item{P4}{probability of YES answer to unrelated question on Q4}
\item{Age}{age in years}
\item{Gender}{gender with categories "Male", "Female" and "Other"}
\item{Difficult}{perceived difficulty of ECWM instructions on a 1-100 scale}
\item{Time}{time to complete the survey in minutes}
}
}
\keyword{datasets}
