% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecwm.R
\name{prev}
\alias{prev}
\title{Prevalence estimation for ECWM data}
\usage{
prev(data, response, prob_var)
}
\arguments{
\item{data}{data frame with individual records containing the
numeric \code{response} the \code{prob_var} variables.}

\item{response}{name of the response variable,
with value 1 denoting "One Yes answer" and value
2 denoting "Two No or two Yes answers".}

\item{prob_var}{name of the variable with the probabilities
of answering "Yes" to the unrelated question.}
}
\value{
A list with the elements:
\describe{
\item{est}{data frame with point and interval estimates for the MME and MLE.}
\item{LR}{likelihood-ratio test against the model with separate
prevalence estimates for each sub-sample.}
\item{freqs}{data frame with the observed and estimated response frequencies.}
}
}
\description{
Point and interval estimates of the
Method of Moments Estimator (MME) and the
Maximum Likelihood Estimator (MLE) for data collected
with the extended crosswise model.
}
\details{
When the MME is outside the parameter space, a
non-parametric bootstrap percentile confidence interval is
reported for the MLE.
}
\examples{
prev(data = doping, response = Q1, prob_var = P1)

}
